move storage location, and make customizable
This changes the location in which Grist files are stored. Tested
under a small test harness docker image rather than on ec2 itself,
so apologies in advance if I missed something.

It is a bit silly since in order to pick up the custom location,
I needed to hardcode the installation path into more scripts.
Feels like there could be a cleaner way to do all this.

Includes a few fixes to small typos.