traefik: enforce TLS by default
It turns out that Authelia needs TLS to work at all, so now that's a
hard requirement. There will be no http-only connexions, all will be
https.

This means that we need to allow self-signed certificates as a last
resort.

Additionally, we also follow grist omnibus in setting an
`HTTPS_METHOD` envvar to either manual or auto, the latter using Let's
Encrypt to obtain its TLS certificate.