traefik: add a health check sidecar service
During startup Dex will need to check up on Traefik that will need to
have a working TLS cert. But Traefik can take a while to do so. So we
should wait for this cert.

We create a new sidecar service to do a health check on traefik based
on the existence of non-default certificate and make Authelia depend
on this service so that we have a nice, sequential, predictable
bootup routine.