packer: clean up the ssh keys
The docs say this is purely cosmetic, because the keys are deleted
from the host (i.e. the machine we ran `packer build .` from), but
they still show up in the guest (the virtual machine image we are
building).

  https://developer.hashicorp.com/packer/integrations/hashicorp/amazon/latest/components/builder/ebs#communicator-configuration

The AWS image checker complains about these keys being there, so let's
be explicit about cleaning them up.