setup-grist-dist: remove docker pull
Pre-populating the image with Docker images would be helpful if we
kept up with the latest release, but so far we have not.

Better to let the user pick the latest versions of all containers when
they first startup Grist instead of baking outdated versions into the
base machine image.