redis: enable new service (#22)
* redis: enable new service

The service is available for both the default and advanced profile.

I picked the default recommendation from the Docker Hub description of
the Redis image.

This also enables memory overcommit in Linux as recommended for Redis
persistence.

In order to allow users to supply their own Redis, we use Docker
Compose profiles to conditionally enable Redis (enabled by default),
and we make the `REDIS_URL` variable configurable.

Co-authored-by: George Gevoian <85144792+georgegevoian@users.noreply.github.com>